add get_all_artists method

+++ spotipy/music/song.py
@@ -1,5 +1,5 @@
 class Song:
-    def __init__(self, id: str, name: str, popularity: float):
+    def __init__(self, id: str, name: str, popularity: int):
         self.id = id
         self.name = name
         self.popularity = popularity
